[MRG] Automatic merge, trunk 18511 to 19725
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/experimental@19726 b456876b-0849-0410-b77d-98878d47e9d5

+++ ws/doc/devtools/tikirelease.sh
@@ -1,6 +1,10 @@
-#!/bin/sh
+#!/bin/sh -x
 # $Id$
 #
+# This script does an export of a given subversion tree and creates a set of
+# release packages that can be uploaded to SourceForge and installed by end 
+# users.
+
 # ==========================================================================
 # IMPORTANT NOTE : This script must NOT be called directly !!!
 #   It is used by the main release script (doc/devtools/release.php)
@@ -9,7 +13,7 @@
 #
 
 SVNROOT="https://tikiwiki.svn.sourceforge.net/svnroot/tikiwiki"
-WORKDIR="/home/$USER/tikipack"
+WORKDIR="$HOME/tikipack"
 MODULE="tikiwiki"
 
 # end of configuration
@@ -22,40 +26,53 @@ echo "Usage: tikirelease.sh <release-version> <svn-relative-path>"
 	exit 0
 fi
 
-OLDIR=`pwd`
+OLDIR="`pwd`"
 VER=$1
 RELTAG=$2
 
 # ############################################################
 
-mkdir -p $WORKDIR
-cd $WORKDIR
+if [ ! -d $WORKDIR ]; then
+    mkdir -p $WORKDIR || die "Can't make $WORKDIR - $!"
+fi
+
+cd $WORKDIR || die "Can't get into $WORKDIR - $!"
+echo "Working in $WORKDIR"
 
 if [ -d $VER ]; then
-	rm -rf $VER
+    echo "Deleting old $VER"
+    rm -rf $VER
 fi
 mkdir $VER
 cd $VER
+
+echo "Exporting $SVNROOT/$RELTAG $MODULE-$VER"
 svn export $SVNROOT/$RELTAG $MODULE-$VER
+
+echo "Cleaning up"
 find $MODULE-$VER -name .cvsignore -type f -exec rm -f {} \;
 find $MODULE-$VER -name .svnignore -type f -exec rm -f {} \;
-find $MODULE-$VER -type d -exec chmod 775 {} \;
-find $MODULE-$VER -type f -exec chmod 664 {} \;
 
-# some more cleanup
 rm -rf $MODULE-$VER/tests
 rm -rf $MODULE-$VER/db/convertscripts
 rm -rf $MODULE-$VER/doc/devtools
 
+echo "Setting permissions"
+find $MODULE-$VER -type d -exec chmod 775 {} \;
+find $MODULE-$VER -type f -exec chmod 664 {} \;
+
+echo "Creating tarballs"
 tar -czf $MODULE-$VER.tar.gz $MODULE-$VER
 tar -cjf $MODULE-$VER.tar.bz2 $MODULE-$VER
 zip -r $MODULE-$VER.zip $MODULE-$VER
+
+ls $WORKDIR/$VER
 
 echo ""
 echo "To upload the 'tarballs', copy-paste and execute the following line (and change '\$SF_LOGIN' by your SF.net login):"
 echo "cd $WORKDIR/$VER; scp $MODULE-$VER.tar.gz $MODULE-$VER.tar.bz2 $MODULE-$VER.zip \$SF_LOGIN@frs.sourceforge.net:uploads"
 echo ""
 
-cd $OLDIR
+cd "$OLDIR"
 
 echo "Done."
